Move rings texture loading to own file
