+++ src/ts/frontend/assets/textures/rings.ts
@@ -1,0 +1,43 @@
+//  This file is part of Cosmos Journeyer
+//
+//  Copyright (C) 2024 Barthélemy Paléologue <barth.paleologue@cosmosjourneyer.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import { Scene } from "@babylonjs/core";
+import { Texture } from "@babylonjs/core/Materials/Textures/texture";
+
+import { ILoadingProgressMonitor } from "../loadingProgressMonitor";
+import { loadTextureAsync } from "./utils";
+
+import saturnRingsPath from "@assets/sol/textures/saturn_rings.png";
+import uranusRingsPath from "@assets/sol/textures/uranus_rings.png";
+
+export type RingsTextures = {
+    saturn: Texture;
+    uranus: Texture;
+};
+
+export async function loadRingsTextures(
+    scene: Scene,
+    progressMonitor: ILoadingProgressMonitor | null,
+): Promise<RingsTextures> {
+    const saturnRingsTexturePromise = loadTextureAsync("SaturnRingsTexture", saturnRingsPath, scene, progressMonitor);
+    const uranusRingsTexturePromise = loadTextureAsync("UranusRingsTexture", uranusRingsPath, scene, progressMonitor);
+
+    return {
+        saturn: await saturnRingsTexturePromise,
+        uranus: await uranusRingsTexturePromise,
+    };
+}
